Added player names and game data export
In this i made sure that users can enter their names at the start and also the player names are shown in prompts and win messages, but "X" and "O" are still used internally. The game moves are stored in a list and exported to a text file game_log.txt after someone wins.